+++ Cemapa/Models/TB_SINCRONIZACAO_SKYHUB.cs
@@ -14,10 +14,10 @@ namespace Cemapa.Models
     
     public partial class TB_SINCRONIZACAO_SKYHUB
     {
-        public long COD_SINCRONIZACAO_SKYHUB { get; set; }
+        public int COD_SINCRONIZACAO_SKYHUB { get; set; }
         public string TIPO_ACAO { get; set; }
         public string IND_SINCRONIZADO { get; set; }
-        public long COD_PRODUTO { get; set; }
+        public int COD_PRODUTO { get; set; }
         public Nullable<System.DateTime> DT_REGISTRO { get; set; }
         public Nullable<System.DateTime> DT_SINCRONIZACAO { get; set; }
